Merge dev/oratk-64-2 into rel/oratk64-2 (#140)
* feat: Add automation for modifying Oracle patch manifests

- Implemented a Python script (`modify_patches.py`) to read, process, and insert patch data into a target YAML file.
- Created a GitHub Actions workflow (`modify_patches.yml`) to automate the execution of the patch modification script on pull requests to the master branch.
- Added unit tests (`modify_patchlist_test.py`) to validate the functionality of the patch modification logic.
- Included a requirements file (`requirements.txt`) for necessary Python dependencies.
- Updated `.gitignore` to exclude Python cache files.
- Added a sample input YAML file (`modify_patchlist.yml`) to serve as a template for patch data.
- Documented the structure and purpose of the patch manifest in `modify_patchlist.md`.
- Cleaned up existing patch entries in `roles/common/defaults/main.yml` to prepare for new entries.

* feat: Add user guide for patch automation process

* fix: Update file extension references from .yaml to .yml in documentation and scripts

* Refactor the GCS download/transfer roles into a more logical and better flowing set of tasks

* Refined quoting for ssh remote commands in Ansible tasks; Minor adjustments to GCS software provisioning tasks

* Set delete_control_node to False for the presubmit test (#320)

* Set delete_control_node to False for the presubmit test.

* increase loop timeout from 2 to 3 hours

* Add `whenever sqlerror exit sql.sqlcode` to all sqlplus blocks

* Set `db_create_file_dest` when not using ASM storage

* Change from `whenever sqlerror exit sql.sqlcode` to `whenever sqlerror exit failure`

* Configure RAC single-instance presubmit tests to use Oracle Multitenant

To get a bit more test coverage, here we configure the RAC single-instance test case (which is the faster one anyway) to set `--ora-db-container` to true, setting up multitenant.

* Fix multitenant check when saving PDB state (#326)

* Fix multitenant check when saving PDB state

* Change `container_db` to a boolean

* Set pwd_gen_cmd conditionally based on ora-db-container flag (#327)

* Set pwd_gen_cmd conditionally based on ora-db-container flag

* remove explicit cast to bool

* Remove `whenever sqlerror` when shutting down from mount; Use PL/SQL to enable flashback and force logging to avoid errors

* Take `oracle_edition` into consideration when building the RDBMS software file list

* Add `list` filter when building `rdbms_edition_sw` for Ansible 2.9 compatibility

* Add new readiness check to ensure that specified storage parameter disk groups exist in `asm_disks`

* Initial commit

* Fix basic syntax errors

* Refactor commit trap and log setup

Modify the common lib to also set up the cleanup trap as part of it's
init, as one less thing the callers might forget.

And talking about forgetting, we had a setup_logging function, and never
called it.  Refactoring so that watch_logs does both the setup and
watching.  It's long, but there's really no need for callers to run part
of it.

* Refactor commit trap and log setup

Modify the common lib to also set up the cleanup trap as part of it's
init, as one less thing the callers might forget.

And talking about forgetting, we had a setup_logging function, and never
called it.  Refactoring so that watch_logs does both the setup and
watching.  It's long, but there's really no need for callers to run part
of it.

* Initial commit

* Fix basic syntax errors

* Clean up merge artifacts

* Update templateing logic

Modify the sed logic to make replacements into a remplrary file and to
use that, allowing the presubmimt to be run multiple times without
clobbering itself.

Also removing a mistaken commented-out apk, adding an error check, and
sending error messages to stderr (&2).

* Add ora_db_container and improve error checks

Free edition runs Oracle 23c, which no longer supports non-CDBs.  Using
the default terraform `ora_db_container` value of false, installs fail.
This change explicitly sets it to `false` for data guard and `true` for
free edition tests.

Also some improved error checking:
* If the sourcing fails, bail out
* If the infra manager apply command fails, run a describe command to
  try and get the actual Terraform error if possible

* Add ora_db_container and improve error checks

Free edition runs Oracle 23c, which no longer supports non-CDBs.  Using
the default terraform `ora_db_container` value of false, installs fail.
This change explicitly sets it to `false` for data guard and `true` for
free edition tests.

Also some improved error checking:
* If the sourcing fails, bail out
* If the infra manager apply command fails, run a describe command to
  try and get the actual Terraform error if possible

* Override and set `ORA_DB_CONTAINER=TRUE` for Free edition installs

* set workload_agent_username conditionally

* Fix DG setup issues related to the staticConnectidentifier value and listener.ora & tnsnames.ora entries

* Use `ansible_hostname` instead of `ansible_fqdn` when building tnsnames.ora entry

* Build and set proper Data Guard StaticConnectIdentifier string

* Changes to support installing Free edition on EL9 with new `os_version` key-value in `rdbms_software`

* Updated filters for building list of RDBMS software to make `os_version` optional

* Make disks larger for free edition test

Oracle 23c's DBCA requires at least 13GB of space on the RECO disk.
Also extending the data disk because, until #335 lands, the DATA and
RECO disks are classified as ASM disks and ignored, so the full install
actually goes on the boot disk.

Separately, make the free edition script executable in case someone
wants to run it directly using ./

* Address review comments

Check if ${deployment_name} is set in setup_vars()
Set `delete-confrol-node` to `false` to avoid a race condition on
long-running presubmits.

* Data Guard changes required to support deployments with XFS storage

* Change tnsnames.ora entries from `LISTENER_{{ db_name }}` to `LISTENER_{{ oracle_sid }}`

* Break `roles/common/defaults/main.yml` into multiple separate files for manageability

* Changes to support optionally procuring software from a URL

* Remove `url_username` and `url_password` for `uri` module tasks

* Refactor output handling in modify_patches.py to create multiple YAML files for different patch types and update tests accordingly

* fix: merge errors

* fix: merge errors

* fix: merge error

* fix: merge error

---------

Co-authored-by: Simon Pane <simonpane@users.noreply.github.com>
Co-authored-by: Simon Pane <pane@pythian.com>
Co-authored-by: AlexBasinov <AlexBasinov@users.noreply.github.com>
Co-authored-by: Marc Fielding <mfielding@google.com>
Co-authored-by: Alexey Basinov <abasinov@google.com>